requirements.txt
pydeseq2>=0.5.0
pyyaml>=6.0.1
seaborn>=0.13.2